MAT-6608 Created util file for better testing in TestCasepopulationList component
