+++ src/components/editTestCase/qdm/populations/TestCasePopulationList.test.tsx
@@ -1,16 +1,8 @@
 import * as React from "react";
 import { render, screen, waitFor } from "@testing-library/react";
 import { MemoryRouter } from "react-router-dom";
-import TestCasePopulationList, {
-  determineGroupResult,
-  determineGroupResultStratification,
-} from "./TestCasePopulationList";
-import {
-  DisplayPopulationValue,
-  DisplayStratificationValue,
-  MeasureScoring,
-  PopulationType,
-} from "@madie/madie-models";
+import TestCasePopulationList from "./TestCasePopulationList";
+import { MeasureScoring, PopulationType } from "@madie/madie-models";
 import userEvent from "@testing-library/user-event";
 
 describe("TestCasePopulationList component", () => {
@@ -39,7 +31,7 @@ describe("TestCasePopulationList component", () => {
           populationBasis="boolean"
           populations={testCasePopulations}
           content="population"
-          i={0}
+          groupIndex={0}
           scoring="Proportion"
         />
       </MemoryRouter>
@@ -94,7 +86,7 @@ describe("TestCasePopulationList component", () => {
           disableExpected={false}
           populationBasis="true"
           content="population"
-          i={0}
+          groupIndex={0}
           scoring="Proportion"
         />
       </MemoryRouter>
@@ -125,14 +117,12 @@ describe("TestCasePopulationList component", () => {
         actual: true,
       },
     ];
-    const testCaseStratifications = [
-      {
-        id: "1",
-        name: "Strata-1 Initial Population",
-        expected: true,
-        actual: true,
-      },
-    ];
+    const testCaseStratification = {
+      id: "1",
+      name: "Strata-1 Initial Population",
+      expected: true,
+      actual: true,
+    };
     const handleChange = jest.fn();
     const handleStratificationChange = jest.fn();
 
@@ -140,13 +130,13 @@ describe("TestCasePopulationList component", () => {
       <MemoryRouter>
         <TestCasePopulationList
           populations={testCasePopulations}
-          stratifications={testCaseStratifications}
+          stratification={testCaseStratification}
           onChange={handleChange}
           onStratificationChange={handleStratificationChange}
           disableExpected={false}
           populationBasis="true"
           content="population"
-          i={0}
+          groupIndex={0}
           scoring="Proportion"
         />
       </MemoryRouter>
@@ -176,14 +166,12 @@ describe("TestCasePopulationList component", () => {
         actual: true,
       },
     ];
-    const testCaseStratifications = [
-      {
-        id: "1",
-        name: "Strata-1 Initial Population",
-        expected: true,
-        actual: true,
-      },
-    ];
+    const testCaseStratification = {
+      id: "1",
+      name: "Strata-1 Initial Population",
+      expected: true,
+      actual: true,
+    };
     const handleChange = jest.fn();
     const handleStratificationChange = jest.fn();
 
@@ -191,15 +179,15 @@ describe("TestCasePopulationList component", () => {
       <MemoryRouter>
         <TestCasePopulationList
           populations={testCasePopulations}
-          stratifications={testCaseStratifications}
+          stratification={testCaseStratification}
           onChange={handleChange}
           onStratificationChange={handleStratificationChange}
           disableExpected={false}
           populationBasis="true"
           content="population"
-          i={0}
+          groupIndex={0}
           scoring="Proportion"
-          executionRun={true}
+          isTestCaseExecuted={true}
         />
       </MemoryRouter>
     );
@@ -273,7 +261,7 @@ describe("TestCasePopulationList component", () => {
           disableExpected={false}
           populationBasis="boolean"
           content="ratio"
-          i={0}
+          groupIndex={0}
           scoring={MeasureScoring.RATIO}
         />
       </MemoryRouter>
@@ -349,7 +337,7 @@ describe("TestCasePopulationList component", () => {
           disableExpected={false}
           populationBasis="boolean"
           content="ratio"
-          i={0}
+          groupIndex={0}
           scoring={MeasureScoring.RATIO}
         />
       </MemoryRouter>
@@ -403,7 +391,7 @@ describe("TestCasePopulationList component", () => {
           disableExpected={false}
           populationBasis="boolean"
           content="ratio"
-          i={0}
+          groupIndex={0}
           scoring={MeasureScoring.CONTINUOUS_VARIABLE}
         />
       </MemoryRouter>
@@ -415,302 +403,5 @@ describe("TestCasePopulationList component", () => {
     expect(tableRows[1]).toHaveTextContent("Measure Population");
     expect(tableRows[2]).toHaveTextContent("Measure Observation 1");
     expect(tableRows[3]).toHaveTextContent("Measure Observation 2");
-  });
-});
-
-describe("determineGroupResult", () => {
-  it("should return initial for missing executionRun input", () => {
-    const output = determineGroupResult("boolean", [], undefined);
-    expect(output).toEqual("initial");
-  });
-
-  it("should return initial for null executionRun input", () => {
-    const output = determineGroupResult("boolean", [], null);
-    expect(output).toEqual("initial");
-  });
-
-  it("should return initial for false executionRun input", () => {
-    const output = determineGroupResult("boolean", [], false);
-    expect(output).toEqual("initial");
-  });
-
-  it("should return pass for empty populations for boolean PopBasis", () => {
-    const output = determineGroupResult("boolean", [], true);
-    expect(output).toEqual("pass");
-  });
-
-  it("should return pass for empty populations for non-boolean PopBasis", () => {
-    const output = determineGroupResult("Encounter", [], true);
-    expect(output).toEqual("pass");
-  });
-
-  it("should return pass for matching populations values (true) for boolean PopBasis", () => {
-    const populations: DisplayPopulationValue[] = [
-      {
-        id: "1",
-        name: PopulationType.INITIAL_POPULATION,
-        actual: true,
-        expected: true,
-      },
-      {
-        id: "2",
-        name: PopulationType.DENOMINATOR,
-        actual: true,
-        expected: true,
-      },
-      {
-        id: "3",
-        name: PopulationType.NUMERATOR,
-        actual: true,
-        expected: true,
-      },
-    ];
-    const output = determineGroupResult("Encounter", populations, true);
-    expect(output).toEqual("pass");
-  });
-
-  it("should return pass for matching populations values (false) for boolean PopBasis", () => {
-    const populations: DisplayPopulationValue[] = [
-      {
-        id: "1",
-        name: PopulationType.INITIAL_POPULATION,
-        actual: false,
-        expected: false,
-      },
-      {
-        id: "2",
-        name: PopulationType.DENOMINATOR,
-        actual: false,
-        expected: false,
-      },
-      {
-        id: "3",
-        name: PopulationType.NUMERATOR,
-        actual: false,
-        expected: false,
-      },
-    ];
-    const output = determineGroupResult("Encounter", populations, true);
-    expect(output).toEqual("pass");
-  });
-
-  it("should return pass for matching populations values (false) for boolean PopBasis", () => {
-    const populations: DisplayPopulationValue[] = [
-      {
-        id: "1",
-        name: PopulationType.INITIAL_POPULATION,
-        actual: false,
-        expected: undefined,
-      },
-      {
-        id: "2",
-        name: PopulationType.DENOMINATOR,
-        actual: false,
-        expected: false,
-      },
-      {
-        id: "3",
-        name: PopulationType.NUMERATOR,
-        actual: false,
-        expected: undefined,
-      },
-    ];
-    const output = determineGroupResult("Encounter", populations, true);
-    expect(output).toEqual("pass");
-  });
-
-  it("should return fail for non-matching populations values for boolean PopBasis", () => {
-    const populations: DisplayPopulationValue[] = [
-      {
-        id: "1",
-        name: PopulationType.INITIAL_POPULATION,
-        actual: true,
-        expected: false,
-      },
-      {
-        id: "2",
-        name: PopulationType.DENOMINATOR,
-        actual: false,
-        expected: false,
-      },
-      {
-        id: "3",
-        name: PopulationType.NUMERATOR,
-        actual: false,
-        expected: undefined,
-      },
-    ];
-    const output = determineGroupResult("Encounter", populations, true);
-    expect(output).toEqual("fail");
-  });
-
-  it("should return fail for non-matching populations values for boolean PopBasis", () => {
-    const populations: DisplayPopulationValue[] = [
-      {
-        id: "1",
-        name: PopulationType.INITIAL_POPULATION,
-        actual: true,
-        expected: true,
-      },
-      {
-        id: "2",
-        name: PopulationType.DENOMINATOR,
-        actual: false,
-        expected: false,
-      },
-      {
-        id: "3",
-        name: PopulationType.NUMERATOR,
-        actual: true,
-        expected: false,
-      },
-    ];
-    const output = determineGroupResult("Encounter", populations, true);
-    expect(output).toEqual("fail");
-  });
-
-  it("should return pass for matching populations values for non-boolean PopBasis", () => {
-    const populations: DisplayPopulationValue[] = [
-      {
-        id: "1",
-        name: PopulationType.INITIAL_POPULATION,
-        actual: 2,
-        expected: 2,
-      },
-      {
-        id: "2",
-        name: PopulationType.DENOMINATOR,
-        actual: 0,
-        expected: 0,
-      },
-      {
-        id: "3",
-        name: PopulationType.NUMERATOR,
-        actual: 1,
-        expected: 1,
-      },
-    ];
-    const output = determineGroupResult("Encounter", populations, true);
-    expect(output).toEqual("pass");
-  });
-
-  it("should return pass for matching populations values (undefined) for non-boolean PopBasis", () => {
-    const populations: DisplayPopulationValue[] = [
-      {
-        id: "1",
-        name: PopulationType.INITIAL_POPULATION,
-        actual: 2,
-        expected: 2,
-      },
-      {
-        id: "2",
-        name: PopulationType.DENOMINATOR,
-        actual: 0,
-        expected: undefined,
-      },
-      {
-        id: "3",
-        name: PopulationType.NUMERATOR,
-        actual: 0,
-        expected: undefined,
-      },
-    ];
-    const output = determineGroupResult("Encounter", populations, true);
-    expect(output).toEqual("pass");
-  });
-
-  it("should return pass for matching populations values (empty) for non-boolean PopBasis", () => {
-    const populations: DisplayPopulationValue[] = [
-      {
-        id: "1",
-        name: PopulationType.INITIAL_POPULATION,
-        actual: 2,
-        expected: 2,
-      },
-      {
-        id: "2",
-        name: PopulationType.DENOMINATOR,
-        actual: 0,
-        expected: null,
-      },
-      {
-        id: "3",
-        name: PopulationType.NUMERATOR,
-        actual: 0,
-        expected: "" as any,
-      },
-    ];
-    const output = determineGroupResult("Encounter", populations, true);
-    expect(output).toEqual("pass");
-  });
-
-  it("should return fail for matching populations values (zero) for non-boolean PopBasis", () => {
-    const populations: DisplayPopulationValue[] = [
-      {
-        id: "1",
-        name: PopulationType.INITIAL_POPULATION,
-        actual: 2,
-        expected: 0,
-      },
-    ];
-    const output = determineGroupResult("Encounter", populations, true);
-    expect(output).toEqual("fail");
-  });
-
-  it("should return fail for matching populations values (undefined) for non-boolean PopBasis", () => {
-    const populations: DisplayPopulationValue[] = [
-      {
-        id: "1",
-        name: PopulationType.INITIAL_POPULATION,
-        actual: 1,
-        expected: undefined,
-      },
-    ];
-    const output = determineGroupResult("Encounter", populations, true);
-    expect(output).toEqual("fail");
-  });
-
-  it("should return fail for matching populations values (empty) for non-boolean PopBasis", () => {
-    const populations: DisplayPopulationValue[] = [
-      {
-        id: "1",
-        name: PopulationType.INITIAL_POPULATION,
-        actual: 1,
-        expected: "" as any,
-      },
-    ];
-    const output = determineGroupResult("Encounter", populations, true);
-    expect(output).toEqual("fail");
-  });
-
-  it("should return pass for matching populations values (empty) for non-boolean PopBasis", () => {
-    const populations: DisplayPopulationValue[] = [
-      {
-        id: "1",
-        name: PopulationType.INITIAL_POPULATION,
-        actual: 2,
-        expected: 2,
-      },
-      {
-        id: "2",
-        name: PopulationType.DENOMINATOR,
-        actual: 0,
-        expected: null,
-      },
-      {
-        id: "3",
-        name: PopulationType.NUMERATOR,
-        actual: 1,
-        expected: 1,
-      },
-      {
-        id: "4",
-        name: PopulationType.DENOMINATOR_EXCLUSION,
-        actual: "" as any,
-        expected: undefined,
-      },
-    ];
-    const output = determineGroupResult("Encounter", populations, true);
-    expect(output).toEqual("pass");
   });
 });
