+++ src/components/editTestCase/qdm/populations/TestCasePopulationList.tsx
@@ -16,10 +16,14 @@ import {
   faCheckCircle,
   faTimesCircle,
 } from "@fortawesome/free-solid-svg-icons";
+import {
+  determineGroupResult,
+  determineGroupResultStratification,
+} from "./TestCasePopulationListUtil";
 
 export interface TestCasePopulationListProps {
   content: string;
-  i?: number;
+  groupIndex?: number;
   scoring: string;
   populations: DisplayPopulationValue[];
   stratification?: StratificationExpectedValue;
@@ -39,70 +43,6 @@ const StyledIcon = styled(FontAwesomeIcon)(
   ]
 );
 
-// Determines if expected and actual values of all populations matches
-export const determineGroupResult = (
-  populationBasis: string,
-  populations: DisplayPopulationValue[],
-  executionRun?: boolean
-) => {
-  if (!executionRun) {
-    return "initial";
-  }
-  for (let i = 0; i < populations?.length; i++) {
-    const population = populations[i];
-    const { expected, actual } = population;
-    if (populationBasis === "boolean" && expected != actual) {
-      return "fail";
-    } else if (populationBasis !== "boolean") {
-      const expectedNum =
-        _.isNil(expected) ||
-        (typeof expected === "string" && _.isEmpty(expected))
-          ? 0
-          : expected;
-      const actualNum =
-        _.isNil(actual) || (typeof actual === "string" && _.isEmpty(actual))
-          ? 0
-          : actual;
-      if (expectedNum != actualNum) {
-        return "fail";
-      }
-    }
-  }
-  return "pass";
-};
-
-// Determines if stratification expected and actual value matches along with all its populations
-export const determineGroupResultStratification = (
-  populationBasis: string,
-  stratification: StratificationExpectedValue,
-  isTestCaseExecuted?: boolean
-) => {
-  if (!isTestCaseExecuted) {
-    return "initial";
-  }
-  const { expected, actual } = stratification;
-  if (populationBasis === "boolean" && expected != actual) {
-    return "fail";
-  } else if (populationBasis !== "boolean") {
-    const expectedNum =
-      _.isNil(expected) || (typeof expected === "string" && _.isEmpty(expected))
-        ? 0
-        : expected;
-    const actualNum =
-      _.isNil(actual) || (typeof actual === "string" && _.isEmpty(actual))
-        ? 0
-        : actual;
-    if (expectedNum != actualNum) {
-      return "fail";
-    }
-  }
-  return determineGroupResult(
-    populationBasis,
-    stratification?.populationValues,
-    isTestCaseExecuted
-  );
-};
-
 // Test case population table. We need to know if the execution has been
 const TestCasePopulationList = ({
   content,
@@ -115,7 +55,7 @@ const TestCasePopulationList = ({
   onChange,
   onStratificationChange,
   errors,
-  i,
+  groupIndex,
 }: TestCasePopulationListProps) => {
   let measureObservations = [];
   let numeratorObservations = [];
@@ -256,7 +196,7 @@ const TestCasePopulationList = ({
           </span>
           <span
             className="sub-caption"
-            data-testid={`${contentId}-scoring-unit-${i + 1}`}
+            data-testid={`${contentId}-scoring-unit-${groupIndex + 1}`}
           >{` - ${scoring} | ${getPatientBasisLabel()}`}</span>
         </caption>
         <thead>
@@ -272,7 +212,7 @@ const TestCasePopulationList = ({
         <tbody>
           {stratification && (
             <TestCaseStratification
-              index={i}
+              index={groupIndex}
               QDM={true}
               strataCode={stratification?.name}
               stratification={stratification}
